+++ R/RcppExports.R
@@ -25,8 +25,8 @@ Leaf__perturb_stem_b <- function(obj_, stem_b) {
     invisible(.Call('_phylloptim_Leaf__perturb_stem_b', PACKAGE = 'phylloptim', obj_, stem_b))
 }
 
-Leaf__set_traits <- function(obj_, vcmax_25, stem_c, stem_b, psi_crit, root_c, root_b, root_psi_crit, beta2, jmax_25, a, curv_fact_elec_trans, curv_fact_colim, cost_scale_TF24) {
-    invisible(.Call('_phylloptim_Leaf__set_traits', PACKAGE = 'phylloptim', obj_, vcmax_25, stem_c, stem_b, psi_crit, root_c, root_b, root_psi_crit, beta2, jmax_25, a, curv_fact_elec_trans, curv_fact_colim, cost_scale_TF24))
+Leaf__set_traits <- function(obj_, vcmax_25, stem_c, stem_b, psi_crit, root_c, root_b, root_psi_crit, beta2, jmax_25, a, curv_fact_elec_trans, curv_fact_colim, cost_scale_TF24, R_d_25) {
+    invisible(.Call('_phylloptim_Leaf__set_traits', PACKAGE = 'phylloptim', obj_, vcmax_25, stem_c, stem_b, psi_crit, root_c, root_b, root_psi_crit, beta2, jmax_25, a, curv_fact_elec_trans, curv_fact_colim, cost_scale_TF24, R_d_25))
 }
 
 Leaf__proportion_of_conductivity <- function(obj_, psi) {
@@ -459,6 +459,30 @@ Leaf__rd_to_vcmax_ratio___get <- function(obj_) {
 
 Leaf__rd_to_vcmax_ratio___set <- function(obj_, value) {
     invisible(.Call('_phylloptim_Leaf__rd_to_vcmax_ratio___set', PACKAGE = 'phylloptim', obj_, value))
+}
+
+Leaf__R_d_25__get <- function(obj_) {
+    .Call('_phylloptim_Leaf__R_d_25__get', PACKAGE = 'phylloptim', obj_)
+}
+
+Leaf__R_d_25__set <- function(obj_, value) {
+    invisible(.Call('_phylloptim_Leaf__R_d_25__set', PACKAGE = 'phylloptim', obj_, value))
+}
+
+Leaf__rd_q10_intercept___get <- function(obj_) {
+    .Call('_phylloptim_Leaf__rd_q10_intercept___get', PACKAGE = 'phylloptim', obj_)
+}
+
+Leaf__rd_q10_intercept___set <- function(obj_, value) {
+    invisible(.Call('_phylloptim_Leaf__rd_q10_intercept___set', PACKAGE = 'phylloptim', obj_, value))
+}
+
+Leaf__rd_q10_slope___get <- function(obj_) {
+    .Call('_phylloptim_Leaf__rd_q10_slope___get', PACKAGE = 'phylloptim', obj_)
+}
+
+Leaf__rd_q10_slope___set <- function(obj_, value) {
+    invisible(.Call('_phylloptim_Leaf__rd_q10_slope___set', PACKAGE = 'phylloptim', obj_, value))
 }
 
 Leaf__PPFD___get <- function(obj_) {